requirements.txt
Flask==2.3.3
python-telegram-bot==20.6
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1